Restored all existent functionality between CategoryFilter componenets
