+++ client/components/sidePanel/categoryFilter/SelectedCategories.js
@@ -1,27 +1,7 @@
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
-import {fetchSelectedCategories} from '../../../store/'
+import React from 'react'
 import CategoryChips from './CategoryChips'
 
-class SelectedCategories extends Component {
-  componentDidMount() {
-    console.log('THIS.PROPS - SELECTED CSATEGORIES', this.props)
-    if (this.props.userId) {
-      this.props.fetchSelectedCategories(this.props.userId)
-    }
-  }
-
-  render() {
-    const userCategories = this.props.selectedCategories
-
-    if (this.props.selectedCategoriesErrored) {
-      return <p>Sorry! There was an error loading your selected filters</p>
-    }
-
-    if (this.props.selectedCategoriesFetching) {
-      return <p>Loading...</p>
-    }
-
+const SelectedCategories = props => {
     return (
       <div>
         <div>
@@ -29,8 +9,8 @@ class SelectedCategories extends Component {
         </div>
         <div>
           <ul className="list selected-categories">
-            {userCategories &&
-              userCategories.map(category => {
+            {props.selectedCategories &&
+              props.selectedCategories.map(category => {
                 return (
                   <li className="list-items" key={category.id}>
                     <CategoryChips
@@ -44,22 +24,6 @@ class SelectedCategories extends Component {
         </div>
       </div>
     )
-  }
 }
 
-const mapStateToProps = state => {
-  return {
-    userId: state.user.id,
-    selectedCategoriesErrored: state.categories.selectedCategoriesErrored,
-    selectedCategoriesFetching: state.categories.selectedCategoriesFetching,
-    selectedCategories: state.categories.selectedCategories
-  }
-}
-
-const mapDispatchToProps = dispatch => {
-  return {
-    fetchSelectedCategories: userId => dispatch(fetchSelectedCategories(userId))
-  }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(SelectedCategories)
+export default (SelectedCategories)
